BlackRoseCoin 'Lover' (v3.5.2)

+++ src/CryptoNoteCore/UpgradeDetector.h
@@ -55,7 +55,7 @@ namespace CryptoNote {
           m_votingCompleteHeight = UNDEF_HEIGHT;
 
         } else if (m_targetVersion - 1 == m_blockchain.back().bl.majorVersion) {
-          m_votingCompleteHeight = findVotingCompleteHeight(static_cast<uint32_t>(m_blockchain.size()) - 1);
+          m_votingCompleteHeight = findVotingCompleteHeight(m_blockchain.size() - 1);
 
         } else if (m_targetVersion <= m_blockchain.back().bl.majorVersion) {
           auto it = std::lower_bound(m_blockchain.begin(), m_blockchain.end(), m_targetVersion,
@@ -65,7 +65,7 @@ namespace CryptoNote {
             return false;
           }
 
-          uint32_t upgradeHeight = static_cast<uint32_t>(it - m_blockchain.begin());
+          uint32_t upgradeHeight = it - m_blockchain.begin();
           m_votingCompleteHeight = findVotingCompleteHeight(upgradeHeight);
           if (m_votingCompleteHeight == UNDEF_HEIGHT) {
             logger(Logging::ERROR, Logging::BRIGHT_RED) << "Internal error: voting complete height isn't found, upgrade height = " << upgradeHeight;
@@ -152,7 +152,7 @@ namespace CryptoNote {
         }
 
       } else {
-        uint32_t lastBlockHeight = static_cast<uint32_t>(m_blockchain.size()) - 1;
+        uint32_t lastBlockHeight = m_blockchain.size() - 1;
         if (isVotingComplete(lastBlockHeight)) {
           m_votingCompleteHeight = lastBlockHeight;
           logger(Logging::INFO, Logging::BRIGHT_GREEN) << "###### UPGRADE voting complete at block index " << m_votingCompleteHeight <<
@@ -193,7 +193,7 @@ namespace CryptoNote {
       assert(m_currency.upgradeHeight(m_targetVersion) == UNDEF_HEIGHT);
 
       uint32_t probableVotingCompleteHeight = probableUpgradeHeight > m_currency.maxUpgradeDistance() ? probableUpgradeHeight - m_currency.maxUpgradeDistance() : 0;
-      for (uint32_t i = probableVotingCompleteHeight; i <= probableUpgradeHeight; ++i) {
+      for (size_t i = probableVotingCompleteHeight; i <= probableUpgradeHeight; ++i) {
         if (isVotingComplete(i)) {
           return i;
         }
